backend/requirements.txt
django
django-cors-headers
requests
python-dotenv
langchain
langchain-openai
langgraph
